move AssetEvent to airflow.sdk.types
